rearranged folders, more button state changes, general fixes
