--- res/mapinfo.txt
@@ -1,7 +1,0 @@
-mvm_bigrock, [wave_start_relay, wave_start_relay_boss, wave_start_relay_ironman, wave_finished_relay], [boss_spawn_relay, boss_door_relay_close, boss_door_relay_open], [boss_dead_relay, boss_deploy_relay], [spawnbot], , [spawnbot_mission_sniper], [spawnbot_mission_spy], [boss_path_a1], [nav_prefer_flank_right, nav_prefer_flank_left, special_main_right, special_main_left], true
-mvm_coaltown, [wave_start_relay, wave_finished_relay], [boss_spawn_relay], [boss_dead_relay, boss_deploy_relay], [spawnbot, spawnbot_giant], , [spawnbot_mission_sniper], [spawnbot_mission_spy], [boss_path_a1], [nav_prefer_flank_right, nav_prefer_flank_left, special_main_right, special_main_left], false
-mvm_decoy, [wave_start_relay, wave_finished_relay], [boss_spawn_relay], [boss_dead_relay, boss_deploy_relay], [spawnbot, spawnbot_invasion, spawnbot_left, spawnbot_right, spawnbot_single_flag], ,[spawnbot_mission_sniper], [spawnbot_mission_spy], [boss_path_a1, boss_path_b1], [nav_prefer_flank_right, nav_prefer_flank_left], true
-mvm_ghost_town, [wave_start_relay, wave_finished_relay, wave_start_relay_666], [boss_spawn_relay], [boss_dead_relay, boss_deploy_relay], [spawnbot, spawnbot_giant], , [spawnbot_mission_sniper], [spawnbot_mission_spy], [boss_path_a1], [nav_prefer_flank_right, nav_prefer_flank_left, special_main_right, special_main_left], false
-mvm_mannhattan, [wave_start_relay, holograms_centerpath_relay, holograms_3way_relay], , , [spawnbot_main0, spawnbot_main1, spawnbot_main1_slow, spawnbot_main2, spawnbot_upper0, spawnbot_upper1, spawnbot_upper2], , [spawnbot_mission_sniper0, spawnbot_mission_sniper1], [spawnbot_mission_spy], , [nav_prefer_gate1_flank, tag_gatebot], true
-mvm_mannworks, [bombpath_arrows_clear_relay, bombpath_arrows_clear_relay_ironman, bombpath_wavefinished], , [boss_deploy_relay], [spawnbot, spawnbot_left, spawnbot_lower, spawnbot_right], , [spawnbot_mission_sniper], [spawnbot_mission_spy], [boss_path_1, boss_path_2_1], [flank_left, flank_right], false
-mvm_rottenburg, [wave_start_relay_classic, wave_start_relay_ironman, wave_finished_relay], [boss_spawn_relay, barricade_destroy_relay, barricade_rebuild_relay], [boss_dead_relay, boss_deploy_relay], [spawnbot, flankers, spawnbot_chief], [spawnbot_mission_sentrybuster], [spawnbot_mission_sniper], [spawnbot_mission_spy], [tank_path_a_10, tank_path_b_10], [flankers], true
